requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
anyio=3.6.2=pyhd8ed1ab_0
appnope=0.1.3=pyhd8ed1ab_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py311he2be06e_3
asttokens=2.2.1=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.12.2=pyha770c72_0
bleach=6.0.0=pyhd8ed1ab_0
bzip2=1.0.8=h3422bc3_4
ca-certificates=2023.5.7=hf0a4a13_0
certifi=2023.5.7=pypi_0
cffi=1.15.1=py311hae827db_3
charset-normalizer=3.1.0=pypi_0
comm=0.1.3=pyhd8ed1ab_0
debugpy=1.6.7=py311ha397e9f_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
filelock=3.12.0=pypi_0
flit-core=3.9.0=pyhd8ed1ab_0
fsspec=2023.5.0=pypi_0
huggingface-hub=0.14.1=pypi_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.6.0=pyha770c72_0
importlib_metadata=6.6.0=hd8ed1ab_0
importlib_resources=5.12.0=pyhd8ed1ab_0
ipykernel=6.23.1=pyh736e0ef_0
ipython=8.13.2=pyhd1c38e8_0
ipython_genutils=0.2.0=py_1
ipywidgets=8.0.6=pyhd8ed1ab_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.2.0=pypi_0
jsonschema=4.17.3=pyhd8ed1ab_0
jupyter_client=8.2.0=pyhd8ed1ab_0
jupyter_core=5.3.0=py311h267d04e_0
jupyter_events=0.6.3=pyhd8ed1ab_0
jupyter_server=2.6.0=pyhd8ed1ab_0
jupyter_server_terminals=0.4.4=pyhd8ed1ab_1
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_widgets=3.0.7=pyhd8ed1ab_1
libcxx=16.0.4=h4653b0c_0
libexpat=2.5.0=hb7217d7_1
libffi=3.4.2=h3422bc3_5
libsodium=1.0.18=h27ca646_1
libsqlite=3.42.0=hb31c410_0
libzlib=1.2.13=h03a7124_4
markupsafe=2.1.2=py311he2be06e_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mistune=2.0.5=pyhd8ed1ab_0
mpmath=1.3.0=pypi_0
nb_conda_kernels=2.3.1=py311h267d04e_2
nbclassic=1.0.0=pyh8b2e9e2_0
nbclient=0.8.0=pyhd8ed1ab_0
nbconvert-core=7.4.0=pyhd8ed1ab_0
nbformat=5.8.0=pyhd8ed1ab_0
ncurses=6.3=h07bb92c_1
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=3.1=pypi_0
notebook=6.5.4=pyha770c72_0
notebook-shim=0.2.3=pyhd8ed1ab_0
numpy=1.24.3=pypi_0
openssl=3.1.0=h53f4e23_3
overrides=7.3.1=pyhd8ed1ab_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.2=pypi_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pip=23.1.2=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
platformdirs=3.5.1=pyhd8ed1ab_0
plotly=5.14.1=pypi_0
prometheus_client=0.17.0=pyhd8ed1ab_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
protobuf=4.23.2=pypi_0
psutil=5.9.5=py311he2be06e_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyaes=1.6.1=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.15.1=pyhd8ed1ab_0
pyobjc-core=9.1.1=py311h490943d_0
pyobjc-framework-cocoa=9.1.1=py311h490943d_0
pyrogram=2.0.106=pypi_0
pyrsistent=0.19.3=py311he2be06e_0
pysocks=1.7.1=pypi_0
python=3.11.3=h1456518_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=1.0.0=pypi_0
python-fastjsonschema=2.17.1=pyhd8ed1ab_0
python-json-logger=2.0.7=pyhd8ed1ab_0
python_abi=3.11=3_cp311
pytz=2023.3=pypi_0
pyyaml=6.0=py311he2be06e_5
pyzmq=25.0.2=py311h0f351f6_0
readline=8.2=h92ec313_1
regex=2023.5.5=pypi_0
requests=2.31.0=pypi_0
rfc3339-validator=0.1.4=pyhd8ed1ab_0
rfc3986-validator=0.1.1=pyh9f0ad1d_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
send2trash=1.8.2=pyhd1c38e8_0
sentencepiece=0.1.99=pypi_0
setuptools=67.7.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pyhd8ed1ab_0
soupsieve=2.3.2.post1=pyhd8ed1ab_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
tenacity=8.2.2=pypi_0
terminado=0.17.1=pyhd1c38e8_0
tgcrypto=1.2.5=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.12=he1e0b03_0
tokenizers=0.13.3=pypi_0
torch=2.0.1=pypi_0
tornado=6.3.2=py311heffc1b2_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
transformers=4.29.2=pypi_0
typing-extensions=4.6.2=hd8ed1ab_0
typing_extensions=4.6.2=pyha770c72_0
typing_utils=0.1.0=pyhd8ed1ab_0
tzdata=2023.3=pypi_0
urllib3=2.0.2=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
webencodings=0.5.1=py_1
websocket-client=1.5.2=pyhd8ed1ab_0
wheel=0.40.0=pyhd8ed1ab_0
widgetsnbextension=4.0.7=pyhd8ed1ab_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
zeromq=4.3.4=hbdafb3b_1
zipp=3.15.0=pyhd8ed1ab_0
